+++ src/Expression/StringCallbackExpression.php
@@ -29,7 +29,7 @@ class StringCallbackExpression extends CallbackExpression
     {
         $this->raw = $raw;
 
-        $this->callback = function (string $value) {
+        $this->callback = function ($value) {
             return strval($value);
         };
 
@@ -41,7 +41,7 @@ class StringCallbackExpression extends CallbackExpression
             $_childCallback = $extractChild->getExpression()->getCallback();
 
             if ($_childCallback) {
-                $this->callback = function (string $value) use ($_childCallback) {
+                $this->callback = function ($value) use ($_childCallback) {
                     return strval((string) $_childCallback($value));
                 };
             }