+++ src/RegisterExpression.php
@@ -12,9 +12,11 @@
 
 namespace Cacing69\Cquery;
 
+use Cacing69\Cquery\Expression\AppendCallbackExpression;
 use Cacing69\Cquery\Expression\AppendNodeCallbackExpression;
 use Cacing69\Cquery\Expression\AttributeCallbackExpression;
 use Cacing69\Cquery\Expression\DefaultCallbackExpression;
+use Cacing69\Cquery\Expression\DoubleQuotesCallbackExpression;
 use Cacing69\Cquery\Expression\FloatCallbackExpression;
 use Cacing69\Cquery\Expression\IntegerCallbackExpression;
 use Cacing69\Cquery\Expression\LengthCallbackExpression;
@@ -22,11 +24,14 @@ use Cacing69\Cquery\Expression\LowerCallbackExpression;
 use Cacing69\Cquery\Expression\ReplaceCallbackExpression;
 use Cacing69\Cquery\Expression\ReverseCallbackExpression;
 use Cacing69\Cquery\Expression\SingleQuotesCallbackExpression;
+use Cacing69\Cquery\Expression\StaticFloatCallbackExpression;
+use Cacing69\Cquery\Expression\StaticIntCallbackExpression;
 use Cacing69\Cquery\Expression\StringCallbackExpression;
 use Cacing69\Cquery\Expression\UpperCallbackExpression;
 
 /**
- * RegisterExpression used to register available Expressions, this Expression is utilized during create definer and filter.
+ * RegisterExpression used to register available Expressions,
+ * this Expression is utilized during create definer and filter.
  *
  * @author Ibnul Mutaki <ibnuu@gmail.com>
  *
@@ -38,18 +43,21 @@ class RegisterExpression
     public static function load()
     {
         return [
+            AppendCallbackExpression::class,
+            AppendNodeCallbackExpression::class,
             AttributeCallbackExpression::class,
+            DoubleQuotesCallbackExpression::class,
+            FloatCallbackExpression::class,
+            IntegerCallbackExpression::class,
             LengthCallbackExpression::class,
             ReverseCallbackExpression::class,
-            UpperCallbackExpression::class,
-            AppendNodeCallbackExpression::class,
-            LowerCallbackExpression::class,
             LowerCallbackExpression::class,
             ReplaceCallbackExpression::class,
             SingleQuotesCallbackExpression::class,
-            IntegerCallbackExpression::class,
+            StaticFloatCallbackExpression::class,
+            StaticIntCallbackExpression::class,
             StringCallbackExpression::class,
-            FloatCallbackExpression::class,
+            UpperCallbackExpression::class,
 
             /**
              *  Make sure that DefaultCallbackExpression is always at the bottom.